# Optimized decimal parsing