--- a/Src/Newtonsoft.Json/Utilities/ConvertUtils.cs
+++ b/Src/Newtonsoft.Json/Utilities/ConvertUtils.cs
@@ -34,6 +34,8 @@ using System.Text;
 using System.Text.RegularExpressions;
 using Newtonsoft.Json.Serialization;
 using System.Reflection;
+using System.Runtime.InteropServices;
+using System.Threading;
 #if NET20
 using Newtonsoft.Json.Utilities.LinqBridge;
 #endif
@@ -1362,6 +1364,256 @@ namespace Newtonsoft.Json.Utilities
 
             value = IEEE754.PackDouble(isNegative, mantissa, exponent);
             return double.IsInfinity(value) ? ParseResult.Overflow : ParseResult.Success;
+        }
+
+        public static ParseResult DecimalTryParse(char[] chars, int start, int length, out decimal value)
+        {
+            value = 0M;
+            const decimal decimalOverflowValue = 7922816251426433759354395033M;
+
+            if (length == 0)
+            {
+                return ParseResult.Invalid;
+            }
+
+            bool isNegative = (chars[start] == '-');
+            if (isNegative)
+            {
+                // text just a negative sign
+                if (length == 1)
+                {
+                    return ParseResult.Invalid;
+                }
+
+                start++;
+                length--;
+            }
+
+            int i = start;
+            int end = start + length;
+            int numDecimalStart = end;
+            int numDecimalEnd = end;
+            int exponent = 0;
+            ulong mantissa = 0UL;
+            int mantissaDigits = 0;
+            int exponentFromMantissa = 0;
+            bool checkForRounding = true;
+            bool roundUp = false;
+            bool? storeOnly28Digits = null;
+            for (; i < end; i++)
+            {
+                char c = chars[i];
+                switch (c)
+                {
+                    case '.':
+                        if (i == start)
+                        {
+                            return ParseResult.Invalid;
+                        }
+                        if (i + 1 == end)
+                        {
+                            return ParseResult.Invalid;
+                        }
+
+                        if (numDecimalStart != end)
+                        {
+                            // multiple decimal points
+                            return ParseResult.Invalid;
+                        }
+
+                        numDecimalStart = i + 1;
+                        break;
+                    case 'e':
+                    case 'E':
+                        if (i == start)
+                        {
+                            return ParseResult.Invalid;
+                        }
+                        if (i == numDecimalStart)
+                        {
+                            // E follows decimal point		
+                            return ParseResult.Invalid;
+                        }
+                        i++;
+                        if (i == end)
+                        {
+                            return ParseResult.Invalid;
+                        }
+
+                        if (numDecimalStart < end)
+                        {
+                            numDecimalEnd = i - 1;
+                        }
+
+                        c = chars[i];
+                        bool exponentNegative = false;
+                        switch (c)
+                        {
+                            case '-':
+                                exponentNegative = true;
+                                i++;
+                                break;
+                            case '+':
+                                i++;
+                                break;
+                        }
+
+                        // parse 3 digit 
+                        for (; i < end; i++)
+                        {
+                            c = chars[i];
+                            if (c < '0' || c > '9')
+                            {
+                                return ParseResult.Invalid;
+                            }
+
+                            int newExponent = (10 * exponent) + (c - '0');
+                            // stops updating exponent when overflowing
+                            if (exponent < newExponent)
+                            {
+                                exponent = newExponent;
+                            }
+                        }
+
+                        if (exponentNegative)
+                        {
+                            exponent = -exponent;
+                        }
+                        break;
+                    default:
+                        if (c < '0' || c > '9')
+                        {
+                            return ParseResult.Invalid;
+                        }
+
+                        if (i == start && c == '0')
+                        {
+                            i++;
+                            if (i != end)
+                            {
+                                c = chars[i];
+                                if (c == '.')
+                                {
+                                    goto case '.';
+                                }
+                                if (c == 'e' || c == 'E')
+                                {
+                                    goto case 'E';
+                                }
+
+                                return ParseResult.Invalid;
+                            }
+                        }
+
+                        if (mantissaDigits < 29 && (mantissaDigits != 28 || !(storeOnly28Digits ?? (storeOnly28Digits = (value > decimalOverflowValue || (value == decimalOverflowValue && c > '5'))).GetValueOrDefault())))
+                        {
+                            if (mantissaDigits < 19)
+                            {
+                                mantissa = (mantissa * 10UL) + (ulong)(c - '0');
+                            }
+                            else 
+                            {
+                                if (mantissaDigits == 19)
+                                {
+                                    value = mantissa;
+                                }
+                                value = (value * 10M) + (c - '0');
+                            }
+                            ++mantissaDigits;
+                        }
+                        else
+                        {
+                            if (checkForRounding)
+                            {
+                                roundUp = c >= '5';
+                                checkForRounding = false;
+                            }
+                            ++exponentFromMantissa;
+                        }
+                        break;
+                }
+            }
+
+            exponent += exponentFromMantissa;
+
+            // correct the decimal point
+            exponent -= (numDecimalEnd - numDecimalStart);
+
+            if (mantissaDigits <= 19)
+            {
+                value = mantissa;
+            }
+
+            if (exponent > 0)
+            {
+                mantissaDigits += exponent;
+                if (mantissaDigits > 29)
+                {
+                    return ParseResult.Overflow;
+                }
+
+                if (exponent > 1)
+                {
+                    value *= DecimalFactors[exponent - 2];
+                }
+                if (mantissaDigits == 29 && value > decimalOverflowValue)
+                {
+                    return ParseResult.Overflow;
+                }
+                value *= 10M;
+            }
+            else
+            {
+                if (roundUp && exponent >= -28)
+                {
+                    ++value;
+                }
+                if (exponent < 0)
+                {
+                    if (mantissaDigits + exponent + 28 <= 0)
+                    {
+                        value = 0M;
+                        return ParseResult.Success; // value is 0
+                    }
+                    if (exponent >= -28)
+                    {
+                        value /= DecimalFactors[-exponent - 1];
+                    }
+                    else
+                    {
+                        value /= DecimalFactors[27];
+                        value /= DecimalFactors[-exponent - 29];
+                    }
+                }
+            }
+
+            if (isNegative)
+            {
+                value *= -1M;
+            }
+
+            return ParseResult.Success;
+        }
+
+        private static decimal[] _decimalFactors;
+
+        private static decimal[] DecimalFactors
+        {
+            get
+            {
+                decimal[] decimalFactors = _decimalFactors;
+                if (decimalFactors == null)
+                {
+                    decimalFactors = new decimal[28];
+                    decimal last = 1M;
+                    for (var i = 0; i < decimalFactors.Length; ++i)
+                    {
+                        decimalFactors[i] = last *= 10M;
+                    }
+                    _decimalFactors = decimalFactors;
+                }
+                return decimalFactors;
+            }
         }
 
         public static bool TryConvertGuid(string s, out Guid g)