Avoids string allocation