@TylerBrinkley 2017-02-06
Using [BenchmarkDotNet](https://github.com/dotnet/BenchmarkDotNet) these are the average times I got deserializing a list of 5000 `decimal`s.

```
x64
  old: 1.8630 ms
  new: 1.2445 ms

x86
  old: 1.7826 ms
  new: 1.3850 ms
```