@JamesNK 2017-02-14
Yeah I'll get to it before 10r1. I've been focusing on Roslyn project stuff lately.