@JamesNK 2017-02-22
I looked over it and the logic and tests look ok. I did a performance test and it was about the same speed to a little slower than the default decimal.TryParse. Any ideas for performance improvements?