@TylerBrinkley 2017-02-22
I just further optimized the parsing performance. I believe that's the best I can get it.