@JamesNK 2017-02-24
I did some more testing, this time with BenchmarkDotNet and yours was faster - I'm not sure whether it was your recent changes or my previous test wasn't that great.

Thanks a lot. Merged!